Add error and confirmation dialog boxes

+++ src/i52salia/aircontrol/Controller.java
@@ -7,6 +7,7 @@ import i52salia.aircontrol.panels.ProgrammingPanel;
 import i52salia.aircontrol.panels.SettingsPanel;
 import i52salia.aircontrol.utils.ACProgram;
 import i52salia.aircontrol.utils.AirConditioner;
+import i52salia.aircontrol.utils.DialogBoxes;
 import i52salia.aircontrol.utils.Temperature;
 import i52salia.aircontrol.utils.Time;
 import java.awt.event.ActionEvent;
@@ -15,6 +16,7 @@ import java.awt.event.MouseEvent;
 import java.util.Locale;
 import java.util.ResourceBundle;
 import javax.swing.DefaultListModel;
+import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.event.ChangeEvent;
 
@@ -24,7 +26,7 @@ import javax.swing.event.ChangeEvent;
  * @author Andrés Salinas Lima (i52salia@uco.es)
  */
 public final class Controller {
-
+    
     private final Model model;
     private View view;
     private AirConditioner selectedDevice;
@@ -39,28 +41,28 @@ public final class Controller {
     public Controller(Model model, View view) {
         this.model = model;
         this.view = view;
-
+        
         initController();
         initView();
     }
-
+    
     private void initController() {
         initTabBarController();
         initHomePanelController();
         initProgrammingPanelController();
         initSettingsPanelController();
     }
-
+    
     private void initView() {
         switchToHomeTab();
-
+        
         switchTemperatureUnit();
         switchTimeFormat();
-
+        
         view.setVisible(true);
         view.setFocusable(true);
     }
-
+    
     private void initTabBarController() {
         view.homeTab.addMouseListener(new MouseAdapter() {
             @Override
@@ -68,14 +70,14 @@ public final class Controller {
                 switchToHomeTab();
             }
         });
-
+        
         view.programmingTab.addMouseListener(new MouseAdapter() {
             @Override
             public void mousePressed(MouseEvent e) {
                 switchToProgrammingTab();
             }
         });
-
+        
         view.settingsTab.addMouseListener(new MouseAdapter() {
             @Override
             public void mousePressed(MouseEvent e) {
@@ -83,43 +85,43 @@ public final class Controller {
             }
         });
     }
-
+    
     private void initHomePanelController() {
         HomePanel hp = view.getHomePanel();
-
+        
         hp.onOffButton.addActionListener((ActionEvent e) -> {
             selectedDevice.setTurnedOn(!selectedDevice.isTurnedOn());
         });
-
+        
         hp.setpointTemperatureSelector.addChangeListener((ChangeEvent e) -> {
             selectedDevice.setSetpointTemp(
                     hp.setpointTemperatureSelector.getSelectedTemperature());
         });
-
+        
         hp.modeButtons.addChangeListener((ChangeEvent e) -> {
             selectedDevice.setMode(hp.modeButtons.getSelectedMode());
         });
-
+        
         hp.fanSpeedSelector.addChangeListener((ChangeEvent e) -> {
             selectedDevice.setFanSpeed(hp.fanSpeedSelector.getSelectedFanSpeed());
         });
-
+        
         hp.backButton.addActionListener((ActionEvent e) -> {
             switchToHomeTab();
         });
     }
-
+    
     private void initProgrammingPanelController() {
         ProgrammingPanel pp = view.getProgrammingPanel();
-
+        
         pp.addProgramButton.addActionListener((ActionEvent e) -> {
             addNewProgramStep1();
         });
-
+        
         pp.cancelChangesButton.addActionListener((ActionEvent e) -> {
             switchToProgrammingTab();
         });
-
+        
         pp.saveChangesButton.addActionListener((ActionEvent e) -> {
             ACProgram modifiedProgram = pp.programSettingsComponent.getSelectedProgram();
             selectedProgram.setEnabled(modifiedProgram.isEnabled());
@@ -128,70 +130,63 @@ public final class Controller {
             selectedProgram.setSetpointTemp(modifiedProgram.getSetpointTemp());
             selectedProgram.setMode(modifiedProgram.getMode());
             selectedProgram.setFanSpeed(modifiedProgram.getFanSpeed());
-
+            
             switchToProgrammingTab();
         });
-
+        
         pp.deleteProgramButton.addActionListener((ActionEvent e) -> {
             confirmSelectedProgramDeletion();
         });
-
+        
         pp.cancelStep1Button.addActionListener((ActionEvent e) -> {
             switchToProgrammingTab();
         });
-
+        
         pp.nextStepButton.addActionListener((ActionEvent e) -> {
             addNewProgramStep2();
         });
-
+        
         pp.backToStep1Button.addActionListener((ActionEvent e) -> {
             addNewProgramStep1();
         });
-
+        
         pp.saveNewProgramButton.addActionListener((ActionEvent e) -> {
             addNewProgramStep3();
         });
     }
-
+    
     private void initSettingsPanelController() {
         SettingsPanel sp = view.getSettingsPanel();
-
+        
         sp.languageComboBox.addActionListener((ActionEvent e) -> {
-            String oldLang = Locale.getDefault().getLanguage();
-
             if (sp.languageComboBox.getSelectedIndex() == 0) {
-                Locale.setDefault(new Locale("en"));
+                switchLanguage("en");
             } else if (sp.languageComboBox.getSelectedIndex() == 1) {
-                Locale.setDefault(new Locale("es"));
+                switchLanguage("es");
             }
-
-            if (!oldLang.equals(Locale.getDefault().getLanguage())) {
-                switchLanguage();
-            }
-
         });
-
+        
         sp.tempUnitComboBox.addActionListener((ActionEvent e) -> {
             if (sp.tempUnitComboBox.getSelectedIndex() == 0) {
                 model.setTempUnit(Temperature.TempUnit.CELSIUS);
             } else if (sp.tempUnitComboBox.getSelectedIndex() == 1) {
                 model.setTempUnit(Temperature.TempUnit.FAHRENHEIT);
             }
-
+            
             switchTemperatureUnit();
         });
-
+        
         sp.timeFormatComboBox.addActionListener((ActionEvent e) -> {
             if (sp.timeFormatComboBox.getSelectedIndex() == 0) {
                 model.setTimeFormat(Time.TimeFormat.TF24HOUR);
             } else if (sp.timeFormatComboBox.getSelectedIndex() == 1) {
                 model.setTimeFormat(Time.TimeFormat.TF12HOUR);
             }
-
+            
             switchTimeFormat();
         });
     }
-
+    
     private void switchToHomeTab() {
         // Prepare home tab
         reloadDeviceList();
@@ -206,7 +201,7 @@ public final class Controller {
         // Set tab title
         view.titleLabel.setText(bundle.getString("View.tabs.Home"));
     }
-
+    
     private void switchToProgrammingTab() {
         // Prepare programming tab
         reloadProgramList();
@@ -223,7 +218,7 @@ public final class Controller {
         // Set tab title
         view.titleLabel.setText(bundle.getString("View.tabs.Programming"));
     }
-
+    
     private void addNewProgramStep1() {
         ProgrammingPanel pp = view.getProgrammingPanel();
 
@@ -246,14 +241,15 @@ public final class Controller {
         // Set tab title
         view.titleLabel.setText(bundle.getString("ProgrammingPanel.NewProgramTitle"));
     }
-
+    
     private void addNewProgramStep2() {
         ProgrammingPanel pp = view.getProgrammingPanel();
 
         // Check if a device has been selected
         int selectedDeviceIndex = pp.devicesList.getSelectedIndex();
         if (selectedDeviceIndex < 0 || selectedDeviceIndex >= model.getDevices().size()) {
-            // TODO: Show warning window
+            DialogBoxes.showErrrorMessage(
+                    view, bundle.getString("Controller.addNewProgramStep2.Error"));
             return;
         }
 
@@ -276,17 +272,17 @@ public final class Controller {
         // Set tab title
         view.titleLabel.setText(bundle.getString("ProgrammingPanel.NewProgramTitle"));
     }
-
+    
     private void addNewProgramStep3() {
         ProgrammingPanel pp = view.getProgrammingPanel();
-
+        
         ACProgram newProgram = pp.newProgramComponent.getSelectedProgram();
-
+        
         selectedDevice.getPrograms().add(newProgram);
-
+        
         switchToProgrammingTab();
     }
-
+    
     private void switchToSettingsTab() {
         SettingsPanel sp = view.getSettingsPanel();
 
@@ -337,12 +333,12 @@ public final class Controller {
         // Set tab title
         view.titleLabel.setText(bundle.getString("View.tabs.Settings"));
     }
-
+    
     private void reloadDeviceList() {
         javax.swing.JPanel listPanel = view.getHomePanel().deviceListPanel;
-
+        
         listPanel.removeAll();
-
+        
         model.getDevices().stream().map((device) -> {
             DeviceListItemComponent newDeviceComponent = new DeviceListItemComponent();
             newDeviceComponent.nameLabel.setText(device.getGivenName());
@@ -385,40 +381,40 @@ public final class Controller {
         }).forEach((newDeviceComponent) -> {
             listPanel.add(newDeviceComponent);
         });
-
+        
         listPanel.revalidate();
         listPanel.repaint();
     }
-
+    
     private void openSelectedDevice() {
         HomePanel hp = view.getHomePanel();
-
+        
         hp.deviceListMainPanel.setVisible(false);
         hp.deviceSettingsMainPanel.setVisible(true);
-
+        
         hp.onOffButton.setTurnedOn(selectedDevice.isTurnedOn());
-
+        
         hp.currentTempLabel.setText(
                 selectedDevice.getCurrentTemp().getString(model.getTempUnit()));
-
+        
         hp.setpointTemperatureSelector.setSelectedTemperature(selectedDevice.getSetpointTemp());
-
+        
         hp.modeButtons.setSelectedMode(selectedDevice.getMode());
-
+        
         hp.fanSpeedSelector.setSelectedFanSpeed(selectedDevice.getFanSpeed());
-
+        
         view.titleLabel.setText(selectedDevice.getGivenName());
     }
-
+    
     private void reloadProgramList() {
         JPanel listPanel = view.getProgrammingPanel().programListPanel;
-
+        
         listPanel.removeAll();
-
+        
         model.getDevices().stream().forEach((device) -> {
             device.getPrograms().stream().map((program) -> {
                 ProgramListItemComponent newDeviceComponent = new ProgramListItemComponent();
-
+                
                 newDeviceComponent.nameLabel.setText(device.getGivenName());
                 newDeviceComponent.toggleButton.setToggledOn(program.isEnabled());
                 newDeviceComponent.daysLabel.setText(
@@ -429,7 +425,7 @@ public final class Controller {
                         program.getSetpointTemp().getString(model.getTempUnit()));
                 newDeviceComponent.modeLabel.setText(program.getModeString());
                 newDeviceComponent.fanSpeedLabel.setText(program.getFanSpeedString());
-
+                
                 newDeviceComponent.addMouseListener(new MouseAdapter() {
                     @Override
                     public void mousePressed(MouseEvent e) {
@@ -441,57 +437,68 @@ public final class Controller {
                 newDeviceComponent.toggleButton.addActionListener((ActionEvent e) -> {
                     program.setEnabled(!program.isEnabled());
                 });
-
+                
                 return newDeviceComponent;
             }).forEach((newDeviceComponent) -> {
                 listPanel.add(newDeviceComponent);
             });
         });
-
+        
         listPanel.revalidate();
         listPanel.repaint();
     }
-
+    
     private void openSelectedProgram() {
         ProgrammingPanel pp = view.getProgrammingPanel();
-
+        
         view.titleLabel.setText(bundle.getString("ProgrammingPanel.ProgramTitle"));
-
+        
         pp.programListMainPanel.setVisible(false);
         pp.programSettingsMainPanel.setVisible(true);
-
+        
         pp.programSettingsComponent.setSelectedDeviceName(selectedDevice.getGivenName());
         pp.programSettingsComponent.setSelectedProgram(selectedProgram);
     }
-
+    
     private void confirmSelectedProgramDeletion() {
-        // TODO: Ask user to confirm
-        deleteSelectedProgram();
+        if (DialogBoxes.confirmDeletion(view)) {
+            deleteSelectedProgram();
+        }
     }
-
+    
     private void deleteSelectedProgram() {
         selectedDevice.getPrograms().remove(selectedProgram);
-
+        
         switchToProgrammingTab();
     }
+    
+    private void switchLanguage(String language) {
+        // Check if it is actually a different language than the current one
+        if (Locale.getDefault().getLanguage().equals(language)) {
+            return; // Nothing to do
+        }
 
-    private void switchLanguage() {
-        // Destroy current window
+        // Set new Locale
+        Locale newLocale = new Locale(language);
+        Locale.setDefault(newLocale);
+        JOptionPane.setDefaultLocale(newLocale); // For predefined dialog boxes
+
+        // Update the language bundle for this controller
+        bundle = ResourceBundle.getBundle(
+                "i52salia/aircontrol/resources/languagebundles/Bundle");
+
+        // Destroy current View (JFrame and all its components)
         view.dispose();
 
-        // Create and initialize new window
+        // Create and initialize a new View
         view = new View();
         initController();
         initView();
 
-        // Update the bundle for this controller
-        bundle = ResourceBundle.getBundle(
-                "i52salia/aircontrol/resources/languagebundles/Bundle");
-
         // Switch back to settings tab
         switchToSettingsTab();
     }
-
+    
     private void switchTemperatureUnit() {
         view.getHomePanel().setpointTemperatureSelector.setTemperatureUnit(
                 model.getTempUnit());
@@ -499,17 +506,17 @@ public final class Controller {
                 model.getTempUnit());
         view.getProgrammingPanel().newProgramComponent.setTemperatureUnit(
                 model.getTempUnit());
-
+        
         reloadDeviceList();
         reloadProgramList();
     }
-
+    
     private void switchTimeFormat() {
         view.getProgrammingPanel().programSettingsComponent.setTimeFormat(
                 model.getTimeFormat());
         view.getProgrammingPanel().newProgramComponent.setTimeFormat(
                 model.getTimeFormat());
-
+        
         reloadDeviceList();
         reloadProgramList();
     }
